Add back evidence as a subcommand and make it fetch input args by itself
