+++ mse_home/command/spawn.py
@@ -1,23 +1,17 @@
 """mse_home.command.spawn module."""
 
 import os
-import socket
-import ssl
 from datetime import datetime, timedelta
 from pathlib import Path
 from uuid import uuid4
 
-from cryptography.hazmat.primitives.serialization import Encoding, load_pem_private_key
-from cryptography.x509 import load_pem_x509_certificate
 from docker.client import DockerClient
 from docker.models.containers import Container
-from intel_sgx_ra.attest import retrieve_collaterals
-from intel_sgx_ra.ratls import get_server_certificate, ratls_verify
 from mse_cli_core.bootstrap import wait_for_conf_server
 from mse_cli_core.clock_tick import ClockTick
-from mse_cli_core.no_sgx_docker import NoSgxDockerConfig
 from mse_cli_core.sgx_docker import SgxDockerConfig
 
+from mse_home.command.evidence import collect_evidence_and_certificate
 from mse_home.command.helpers import (
     app_container_exists,
     get_app_container,
@@ -27,7 +21,6 @@ from mse_home.command.helpers import (
 )
 from mse_home.log import LOGGER as LOG
 from mse_home.model.code import CodeConfig
-from mse_home.model.evidence import ApplicationEvidence
 from mse_home.model.package import CodePackage
 
 
@@ -154,10 +147,9 @@ def run(args) -> None:
     LOG.info("The application is now ready to receive the secrets!")
 
     # Generate evidence and RA-TLS certificate files
-    app_args = NoSgxDockerConfig.from_sgx(docker_config)
     container: Container = get_app_container(client, args.name)
 
-    collect_evidence_and_certificate(container, args.pccs, args.output, app_args)
+    collect_evidence_and_certificate(container, args.pccs, args.output)
 
 
 def run_docker_image(
@@ -187,58 +179,3 @@ def run_docker_image(
         raise Exception(
             f"Can't create the container: {container.status} - {container.logs()}"
         )
-
-
-# pylint: disable=too-many-locals
-def collect_evidence_and_certificate(
-    container: Container,
-    pccs_url: str,
-    output: Path,
-    app_args: NoSgxDockerConfig,
-):
-    """Collect evidence JSON file and RA-TLS certificate from running enclave."""
-    docker = SgxDockerConfig.load(container.attrs, container.labels)
-
-    # Get the certificate from the application
-    try:
-        ratls_cert = load_pem_x509_certificate(
-            get_server_certificate(("localhost", docker.port)).encode("utf-8")
-        )
-    except (ssl.SSLZeroReturnError, socket.gaierror, ssl.SSLEOFError) as exc:
-        raise ConnectionError(
-            f"Can't reach localhost:{docker.port}. "
-            "Are you sure the application is still running?"
-        ) from exc
-
-    quote = ratls_verify(ratls_cert)
-
-    (tcb_info, tcb_cert, root_ca_crl, pck_platform_crl) = retrieve_collaterals(
-        quote, pccs_url
-    )
-
-    signer_key = load_pem_private_key(
-        docker.signer_key.read_bytes(),
-        password=None,
-    )
-
-    evidence = ApplicationEvidence(
-        app_args=app_args.json(),
-        ratls_certificate=ratls_cert,
-        root_ca_crl=root_ca_crl,
-        pck_platform_crl=pck_platform_crl,
-        tcb_info=tcb_info,
-        tcb_cert=tcb_cert,
-        signer_pk=signer_key.public_key(),
-    )
-
-    evidence_path = output / "evidence.json"
-    evidence.save(evidence_path)
-    LOG.info("The evidence file has been generated at: %s", evidence_path)
-    LOG.info("The evidence file can now be shared!")
-
-    ratls_cert_path = output / "ratls.pem"
-    ratls_cert_path.write_bytes(
-        evidence.ratls_certificate.public_bytes(encoding=Encoding.PEM)
-    )
-
-    LOG.info("The RA-TLS certificate has been saved at: %s", ratls_cert_path)
