--- tests/data/args.toml
@@ -1,5 +1,0 @@
-host = "localhost"
-expiration_date = 1714058115
-size = 4096
-app_id = "63322f85-1ff8-4483-91ae-f18d7398d157"
-application = "app:app"
