+++ tests/test_model_evidence.py
@@ -19,7 +19,7 @@ def test_load():
     conf = ApplicationEvidence.load(path=json)
 
     ref_conf = ApplicationEvidence(
-        app_args="{\"host\": \"localhost\", \"expiration_date\": 1714058115, \"app_cert\": null, \"size\": 4096, \"app_id\": \"63322f85-1ff8-4483-91ae-f18d7398d157\", \"application\": \"app:app\"}",
+        input_args="{\"host\": \"localhost\", \"expiration_date\": 1714058115, \"app_cert\": null, \"size\": 4096, \"app_id\": \"63322f85-1ff8-4483-91ae-f18d7398d157\", \"application\": \"app:app\"}",
         ratls_certificate=load_pem_x509_certificate(
             b"-----BEGIN CERTIFICATE-----\nMIIUfDCCFCKgAwIBAgITdFsNOA0J4klfaNLxtAArloOHwjAKBggqhkjOPQQDAjBg\nMQswCQYDVQQGEwJGUjEWMBQGA1UECAwNSWxlLWRlLUZyYW5jZTEOMAwGA1UEBwwF\nUGFyaXMxFTATBgNVBAoMDENvc21pYW4gVGVjaDESMBAGA1UEAwwJbG9jYWxob3N0\nMB4XDTIzMDUxNzA5Mzg0MloXDTI0MDQyNjA5MzgzM1owYDELMAkGA1UEBhMCRlIx\nFjAUBgNVBAgMDUlsZS1kZS1GcmFuY2UxDjAMBgNVBAcMBVBhcmlzMRUwEwYDVQQK\nDAxDb3NtaWFuIFRlY2gxEjAQBgNVBAMMCWxvY2FsaG9zdDBZMBMGByqGSM49AgEG\nCCqGSM49AwEHA0IABJRCBMEF70mXMUT58XTvOlLo9mAr6vAYQvf+gmIAf/v1uFl8\nd+7DKP7aV9jA6jI4rJ1jO1vrMX0dT8AbMft7tA6jghK5MIIStTAUBgNVHREEDTAL\ngglsb2NhbGhvc3QwghKNBgkqhkiG+E2KOQYEghJ+AwACAAAAAAAJAA4Ak5pyM/ec\nTKmUCg2zlX8GB0Juveh/qjdCKYvIfG98s4gAAAAABQsICf//AAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABQAAAAAAAADnAAAAAAAAAA5O\nxV5pCwaldK9Z39UbqjhmrEyNN4lu/8Dx+tF7Mc87AAAAAAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAADBwWHQ3ZluiphH3mfqLAAiZ2H3cVosQi0wEqwQeVoe9QAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEA\nCgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAACNiDYO0OsUm0CiILKp6y8tfeVZUvuLR1nt80Okno34HAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAyhAAAFNDR7kK/DSY+LawSiTH\n7g029V/QQd1tyuMMuo4iX/WOY910502mAbR/WLkmaOAbUFcKo8yD0JREKg6i79Vz\nQv/2ypv/LrT1QmTui7xCGxFpf7ZU+21A28v1PXWJg72lLICVyEkVbQyebjH1+Mzh\n2XOzTex+vsroa7e38lzyMSD0BQsICf//AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAFQAAAAAAAADnAAAAAAAAABkqpQzhwM7wPM+J57Wx\naw15ePXCse3Pd02HcC6BVNi/AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAACMT1d115ZQPpYTf3fGioKaAFasje1wFAsIGwlEkMV7/wAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEACQAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\nAAAUH84PQ7tWUp7et8b37nnxyiQ6noHprp7pWJGn/Y+CBAAAAAAAAAAAAAAAAAAA\nAAAAAAAAAAAAAAAAAAAAAAAAHqKiZMgyhNPxbit5R5fDHqGrSOsaMsbal3fR++xS\nSHBQFgEqse/cmdEgsrVK238LzK4zdldTvKAd2hejemtiUiAAAAECAwQFBgcICQoL\nDA0ODxAREhMUFRYXGBkaGxwdHh8FAGIOAAAtLS0tLUJFR0lOIENFUlRJRklDQVRF\nLS0tLS0KTUlJRThqQ0NCSmlnQXdJQkFnSVVjeTc1bDVBb0FRdjBkOU1TNTNmaGdo\nUGFEV0F3Q2dZSUtvWkl6ajBFQXdJdwpjREVpTUNBR0ExVUVBd3daU1c1MFpXd2dV\nMGRZSUZCRFN5QlFiR0YwWm05eWJTQkRRVEVhTUJnR0ExVUVDZ3dSClNXNTBaV3dn\nUTI5eWNHOXlZWFJwYjI0eEZEQVNCZ05WQkFjTUMxTmhiblJoSUVOc1lYSmhNUXN3\nQ1FZRFZRUUkKREFKRFFURUxNQWtHQTFVRUJoTUNWVk13SGhjTk1qTXdNVEU1TURr\neU5qUXlXaGNOTXpBd01URTVNRGt5TmpReQpXakJ3TVNJd0lBWURWUVFEREJsSmJu\nUmxiQ0JUUjFnZ1VFTkxJRU5sY25ScFptbGpZWFJsTVJvd0dBWURWUVFLCkRCRkpi\nblJsYkNCRGIzSndiM0poZEdsdmJqRVVNQklHQTFVRUJ3d0xVMkZ1ZEdFZ1EyeGhj\nbUV4Q3pBSkJnTlYKQkFnTUFrTkJNUXN3Q1FZRFZRUUdFd0pWVXpCWk1CTUdCeXFH\nU000OUFnRUdDQ3FHU000OUF3RUhBMElBQkhWYQp1T1J1LytaOXBRZDZORnNKQkRr\nbGVSWUFCR1JnRWZ2Z2t2a0k5V3hSTjF6c1g4K0locHhkOThlc3NYK09VRmJ0CmJM\nd3RxZUlnbE43bUxFMXl0b3VqZ2dNT01JSURDakFmQmdOVkhTTUVHREFXZ0JTVmIx\nM052UnZoNlVCSnlkVDAKTTg0QlZ3dmVWREJyQmdOVkhSOEVaREJpTUdDZ1hxQmNo\nbHBvZEhSd2N6b3ZMMkZ3YVM1MGNuVnpkR1ZrYzJWeQpkbWxqWlhNdWFXNTBaV3d1\nWTI5dEwzTm5lQzlqWlhKMGFXWnBZMkYwYVc5dUwzWTBMM0JqYTJOeWJEOWpZVDF3\nCmJHRjBabTl5YlNabGJtTnZaR2x1Wnoxa1pYSXdIUVlEVlIwT0JCWUVGSlZTcmc0\nTnJBazhzNGpXRWFnMVdFU2MKcUNaTU1BNEdBMVVkRHdFQi93UUVBd0lHd0RBTUJn\nTlZIUk1CQWY4RUFqQUFNSUlDT3dZSktvWklodmhOQVEwQgpCSUlDTERDQ0FpZ3dI\nZ1lLS29aSWh2aE5BUTBCQVFRUTBqZ0V3V3RpalpGWmd4MHZOOEJ6WnpDQ0FXVUdD\naXFHClNJYjRUUUVOQVFJd2dnRlZNQkFHQ3lxR1NJYjRUUUVOQVFJQkFnRUVNQkFH\nQ3lxR1NJYjRUUUVOQVFJQ0FnRUUKTUJBR0N5cUdTSWI0VFFFTkFRSURBZ0VETUJB\nR0N5cUdTSWI0VFFFTkFRSUVBZ0VETUJFR0N5cUdTSWI0VFFFTgpBUUlGQWdJQS96\nQVJCZ3NxaGtpRytFMEJEUUVDQmdJQ0FQOHdFQVlMS29aSWh2aE5BUTBCQWdjQ0FR\nQXdFQVlMCktvWklodmhOQVEwQkFnZ0NBUUF3RUFZTEtvWklodmhOQVEwQkFna0NB\nUUF3RUFZTEtvWklodmhOQVEwQkFnb0MKQVFBd0VBWUxLb1pJaHZoTkFRMEJBZ3ND\nQVFBd0VBWUxLb1pJaHZoTkFRMEJBZ3dDQVFBd0VBWUxLb1pJaHZoTgpBUTBCQWcw\nQ0FRQXdFQVlMS29aSWh2aE5BUTBCQWc0Q0FRQXdFQVlMS29aSWh2aE5BUTBCQWc4\nQ0FRQXdFQVlMCktvWklodmhOQVEwQkFoQUNBUUF3RUFZTEtvWklodmhOQVEwQkFo\nRUNBUXN3SHdZTEtvWklodmhOQVEwQkFoSUUKRUFRRUF3UC8vd0FBQUFBQUFBQUFB\nQUF3RUFZS0tvWklodmhOQVEwQkF3UUNBQUF3RkFZS0tvWklodmhOQVEwQgpCQVFH\nTUdCcUFBQUFNQThHQ2lxR1NJYjRUUUVOQVFVS0FRRXdIZ1lLS29aSWh2aE5BUTBC\nQmdRUVRuOCtHMHdYClNzSTJtR25SQVkyQmNUQkVCZ29xaGtpRytFMEJEUUVITURZ\nd0VBWUxLb1pJaHZoTkFRMEJCd0VCQWY4d0VBWUwKS29aSWh2aE5BUTBCQndJQkFR\nQXdFQVlMS29aSWh2aE5BUTBCQndNQkFmOHdDZ1lJS29aSXpqMEVBd0lEU0FBdwpS\nUUloQVArK0VkclFmRmxEM0F2OVU3VmhlT3BCNlNvaDNZSDFPd2ZUVEdCMXloa1BB\naUFEQ1J1K3pWbjhNdTNGCnljMm9nSDNUTnp3Zjh6ZEhkbHNlUE55SUQrRkhEdz09\nCi0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0KLS0tLS1CRUdJTiBDRVJUSUZJQ0FU\nRS0tLS0tCk1JSUNsakNDQWoyZ0F3SUJBZ0lWQUpWdlhjMjlHK0hwUUVuSjFQUXp6\nZ0ZYQzk1VU1Bb0dDQ3FHU000OUJBTUMKTUdneEdqQVlCZ05WQkFNTUVVbHVkR1Zz\nSUZOSFdDQlNiMjkwSUVOQk1Sb3dHQVlEVlFRS0RCRkpiblJsYkNCRApiM0p3YjNK\naGRHbHZiakVVTUJJR0ExVUVCd3dMVTJGdWRHRWdRMnhoY21FeEN6QUpCZ05WQkFn\nTUFrTkJNUXN3CkNRWURWUVFHRXdKVlV6QWVGdzB4T0RBMU1qRXhNRFV3TVRCYUZ3\nMHpNekExTWpFeE1EVXdNVEJhTUhBeElqQWcKQmdOVkJBTU1HVWx1ZEdWc0lGTkhX\nQ0JRUTBzZ1VHeGhkR1p2Y20wZ1EwRXhHakFZQmdOVkJBb01FVWx1ZEdWcwpJRU52\nY25CdmNtRjBhVzl1TVJRd0VnWURWUVFIREF0VFlXNTBZU0JEYkdGeVlURUxNQWtH\nQTFVRUNBd0NRMEV4CkN6QUpCZ05WQkFZVEFsVlRNRmt3RXdZSEtvWkl6ajBDQVFZ\nSUtvWkl6ajBEQVFjRFFnQUVOU0IvN3QyMWxYU08KMkN1enB4dzc0ZUpCNzJFeURH\nZ1c1clhDdHgydFZUTHE2aEtrNnorVWlSWkNucVI3cHNPdmdxRmVTeGxtVGxKbApl\nVG1pMldZejNxT0J1ekNCdURBZkJnTlZIU01FR0RBV2dCUWlaUXpXV3AwMGlmT0R0\nSlZTdjFBYk9TY0dyREJTCkJnTlZIUjhFU3pCSk1FZWdSYUJEaGtGb2RIUndjem92\nTDJObGNuUnBabWxqWVhSbGN5NTBjblZ6ZEdWa2MyVnkKZG1salpYTXVhVzUwWld3\ndVkyOXRMMGx1ZEdWc1UwZFlVbTl2ZEVOQkxtUmxjakFkQmdOVkhRNEVGZ1FVbFc5\nZAp6YjBiNGVsQVNjblU5RFBPQVZjTDNsUXdEZ1lEVlIwUEFRSC9CQVFEQWdFR01C\nSUdBMVVkRXdFQi93UUlNQVlCCkFmOENBUUF3Q2dZSUtvWkl6ajBFQXdJRFJ3QXdS\nQUlnWHNWa2kwdytpNlZZR1czVUYvMjJ1YVhlMFlKRGoxVWUKbkErVGpEMWFpNWND\nSUNZYjFTQW1ENXhrZlRWcHZvNFVveWlTWXhyRFdMbVVSNENJOU5LeWZQTisKLS0t\nLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQotLS0tLUJFR0lOIENFUlRJRklDQVRFLS0t\nLS0KTUlJQ2p6Q0NBalNnQXdJQkFnSVVJbVVNMWxxZE5JbnpnN1NWVXI5UUd6a25C\ncXd3Q2dZSUtvWkl6ajBFQXdJdwphREVhTUJnR0ExVUVBd3dSU1c1MFpXd2dVMGRZ\nSUZKdmIzUWdRMEV4R2pBWUJnTlZCQW9NRVVsdWRHVnNJRU52CmNuQnZjbUYwYVc5\ndU1SUXdFZ1lEVlFRSERBdFRZVzUwWVNCRGJHRnlZVEVMTUFrR0ExVUVDQXdDUTBF\neEN6QUoKQmdOVkJBWVRBbFZUTUI0WERURTRNRFV5TVRFd05EVXhNRm9YRFRRNU1U\nSXpNVEl6TlRrMU9Wb3dhREVhTUJnRwpBMVVFQXd3UlNXNTBaV3dnVTBkWUlGSnZi\nM1FnUTBFeEdqQVlCZ05WQkFvTUVVbHVkR1ZzSUVOdmNuQnZjbUYwCmFXOXVNUlF3\nRWdZRFZRUUhEQXRUWVc1MFlTQkRiR0Z5WVRFTE1Ba0dBMVVFQ0F3Q1EwRXhDekFK\nQmdOVkJBWVQKQWxWVE1Ga3dFd1lIS29aSXpqMENBUVlJS29aSXpqMERBUWNEUWdB\nRUM2bkV3TURJWVpPai9pUFdzQ3phRUtpNwoxT2lPU0xSRmhXR2pibkJWSmZWbmtZ\nNHUzSWprRFlZTDBNeE80bXFzeVlqbEJhbFRWWXhGUDJzSkJLNXpsS09CCnV6Q0J1\nREFmQmdOVkhTTUVHREFXZ0JRaVpReldXcDAwaWZPRHRKVlN2MUFiT1NjR3JEQlNC\nZ05WSFI4RVN6QkoKTUVlZ1JhQkRoa0ZvZEhSd2N6b3ZMMk5sY25ScFptbGpZWFJs\nY3k1MGNuVnpkR1ZrYzJWeWRtbGpaWE11YVc1MApaV3d1WTI5dEwwbHVkR1ZzVTBk\nWVVtOXZkRU5CTG1SbGNqQWRCZ05WSFE0RUZnUVVJbVVNMWxxZE5JbnpnN1NWClVy\nOVFHemtuQnF3d0RnWURWUjBQQVFIL0JBUURBZ0VHTUJJR0ExVWRFd0VCL3dRSU1B\nWUJBZjhDQVFFd0NnWUkKS29aSXpqMEVBd0lEU1FBd1JnSWhBT1cvNVFrUitTOUNp\nU0RjTm9vd0x1UFJMc1dHZi9ZaTdHU1g5NEJnd1R3ZwpBaUVBNEowbHJIb01zK1hv\nNW8vc1g2TzlRV3hIUkF2WlVHT2RSUTdjdnFSWGFxST0KLS0tLS1FTkQgQ0VSVElG\nSUNBVEUtLS0tLQoAMAwGA1UdEwEB/wQCMAAwCgYIKoZIzj0EAwIDSAAwRQIhAI1O\nnFt7sEjvhtWi+G5Ewr1bhLD9mvvHRy24MUwbs08eAiAokkeGBBfREzIs5bU7+PtB\npLMzjW4ikTqPVuAq0bhKzA==\n-----END CERTIFICATE-----\n"
         ),
@@ -40,7 +40,7 @@ def test_load():
         ),
     )
 
-    assert conf.app_args == ref_conf.app_args
+    assert conf.input_args == ref_conf.input_args
     assert conf.ratls_certificate == ref_conf.ratls_certificate
     assert conf.root_ca_crl == ref_conf.root_ca_crl
     assert conf.pck_platform_crl == ref_conf.pck_platform_crl
